Split music into multiple files by implementation
macOS is untested.
