+++ client/sdl/i_musicsystem.h
@@ -1,4 +1,4 @@
-// Emacs style mode select   -*- C++ -*- 
+// Emacs style mode select   -*- C++ -*-
 //-----------------------------------------------------------------------------
 //
 // $Id$
@@ -20,20 +20,9 @@
 //
 //-----------------------------------------------------------------------------
 
-#ifndef __I_MUSICSYSTEM_H__
-#define __I_MUSICSYSTEM_H__
+#pragma once
 
-#include <SDL_mixer.h>
-#include "i_music.h"
 #include "i_midi.h"
-
-#ifdef OSX
-#include <AudioToolbox/AudioToolbox.h>
-#endif	// OSX
-
-#ifdef PORTMIDI
-#include "portmidi.h"
-#endif	// PORTMIDI
 
 // ============================================================================
 //
@@ -46,41 +35,70 @@
 
 class MusicSystem
 {
-public:
-	MusicSystem() : mIsPlaying(false), mIsPaused(false), mTempo(120.0f), mVolume(1.0f) {}
-	virtual ~MusicSystem() {}
-	
+  public:
+	MusicSystem() : mIsPlaying(false), mIsPaused(false), mTempo(120.0f), mVolume(1.0f) { }
+	virtual ~MusicSystem() { }
+
 	virtual void startSong(byte* data, size_t length, bool loop);
 	virtual void stopSong();
 	virtual void pauseSong();
 	virtual void resumeSong();
 	virtual void playChunk() = 0;
-	
-	virtual void setVolume(float volume);
-	float getVolume() const { return mVolume; }
-	virtual void setTempo(float tempo);
-	float getTempo() const { return mTempo; }
-	
-	virtual bool isInitialized() const = 0;
-	bool isPlaying() const { return mIsPlaying; }
-	bool isPaused() const { return mIsPaused; }
-	
-	// Can this MusicSystem object play a particular type of music file?
-	virtual bool isMusCapable() const { return false; }
-	virtual bool isMidiCapable() const { return false; }
-	virtual bool isOggCapable() const { return false; }
-	virtual bool isMp3Capable() const { return false; }
-	virtual bool isModCapable() const { return false; }
-	virtual bool isWaveCapable() const { return false; }
-	
-private:
-	bool	mIsPlaying;
-	bool	mIsPaused;
-	
-	float	mTempo;
-	float	mVolume;
-};
 
+	virtual void setVolume(float volume);
+	float getVolume() const
+	{
+		return mVolume;
+	}
+	virtual void setTempo(float tempo);
+	float getTempo() const
+	{
+		return mTempo;
+	}
+
+	virtual bool isInitialized() const = 0;
+	bool isPlaying() const
+	{
+		return mIsPlaying;
+	}
+	bool isPaused() const
+	{
+		return mIsPaused;
+	}
+
+	// Can this MusicSystem object play a particular type of music file?
+	virtual bool isMusCapable() const
+	{
+		return false;
+	}
+	virtual bool isMidiCapable() const
+	{
+		return false;
+	}
+	virtual bool isOggCapable() const
+	{
+		return false;
+	}
+	virtual bool isMp3Capable() const
+	{
+		return false;
+	}
+	virtual bool isModCapable() const
+	{
+		return false;
+	}
+	virtual bool isWaveCapable() const
+	{
+		return false;
+	}
+
+  private:
+	bool mIsPlaying;
+	bool mIsPaused;
+
+	float mTempo;
+	float mVolume;
+};
 
 // ============================================================================
 //
@@ -93,115 +111,50 @@ private:
 
 class SilentMusicSystem : public MusicSystem
 {
-public:
-	SilentMusicSystem() { Printf(PRINT_WARNING, "I_InitMusic: Music playback disabled.\n"); }
-	
-	virtual void startSong(byte* data, size_t length, bool loop) {}
-	virtual void stopSong() {}
-	virtual void pauseSong() {}
-	virtual void resumeSong() {}
-	virtual void playChunk() {}
-	virtual void setVolume(float volume) const {}
-	
-	virtual bool isInitialized() const { return true; }
-	
+  public:
+	SilentMusicSystem()
+	{
+		Printf(PRINT_WARNING, "I_InitMusic: Music playback disabled.\n");
+	}
+
+	virtual void startSong(byte* data, size_t length, bool loop) { }
+	virtual void stopSong() { }
+	virtual void pauseSong() { }
+	virtual void resumeSong() { }
+	virtual void playChunk() { }
+	virtual void setVolume(float volume) const { }
+
+	virtual bool isInitialized() const
+	{
+		return true;
+	}
+
 	// SilentMusicSystem can handle any type of music by doing nothing
-	virtual bool isMusCapable() const { return true; }
-	virtual bool isMidiCapable() const { return true; }
-	virtual bool isOggCapable() const { return true; }
-	virtual bool isMp3Capable() const { return true; }
-	virtual bool isModCapable() const { return true; }
-	virtual bool isWaveCapable() const { return true; }
+	virtual bool isMusCapable() const
+	{
+		return true;
+	}
+	virtual bool isMidiCapable() const
+	{
+		return true;
+	}
+	virtual bool isOggCapable() const
+	{
+		return true;
+	}
+	virtual bool isMp3Capable() const
+	{
+		return true;
+	}
+	virtual bool isModCapable() const
+	{
+		return true;
+	}
+	virtual bool isWaveCapable() const
+	{
+		return true;
+	}
 };
-
-
-// ============================================================================
-//
-// SdlMixerMusicSystem class
-//
-// Plays music utilizing the SDL_Mixer library and can handle a wide range of
-// music formats.
-//
-// ============================================================================
-
-class SdlMixerMusicSystem : public MusicSystem
-{
-public:
-	SdlMixerMusicSystem();
-	virtual ~SdlMixerMusicSystem();
-	
-	virtual void startSong(byte* data, size_t length, bool loop);
-	virtual void stopSong();
-	virtual void pauseSong();
-	virtual void resumeSong();
-	virtual void playChunk() {}
-	virtual void setVolume(float volume);
-	
-	virtual bool isInitialized() const { return mIsInitialized; }
-
-	virtual bool isMusCapable() const { return true; }
-	virtual bool isMidiCapable() const { return true; }
-	virtual bool isOggCapable() const { return true; }
-	virtual bool isMp3Capable() const { return true; }
-	virtual bool isModCapable() const { return true; }
-	virtual bool isWaveCapable() const { return true; }
-
-private:
-	bool					mIsInitialized;
-	MusicHandler_t			mRegisteredSong;
-
-	void _StopSong();
-	void _RegisterSong(byte* data, size_t length);
-	void _UnregisterSong();
-};
-
-
-// ============================================================================
-//
-// AuMusicSystem class
-//
-// Plays music utilizing OSX's Audio Unit system, which is the default for OSX.
-// On non-OSX systems, the AuMusicSystem will not output any sound and should
-// not be selected.
-//
-// ============================================================================
-
-#ifdef OSX
-class AuMusicSystem : public MusicSystem
-{
-public:
-	AuMusicSystem();
-	virtual ~AuMusicSystem();
-	
-	virtual void startSong(byte* data, size_t length, bool loop);
-	virtual void stopSong();
-	virtual void pauseSong();
-	virtual void resumeSong();
-	virtual void playChunk() {}
-	virtual void setVolume(float volume);
-	
-	virtual bool isInitialized() const { return mIsInitialized; }
-	
-	// Only plays midi-type music
-	virtual bool isMusCapable() const { return true; }
-	virtual bool isMidiCapable() const { return true; }
-	
-private:
-	bool			mIsInitialized;
-	
-	MusicPlayer		mPlayer;
-	MusicSequence	mSequence;
-	AUGraph			mGraph;
-	AUNode			mSynth;
-	AUNode			mOutput;
-	AudioUnit		mUnit;
-	CFDataRef		mCfd;
-		
-	void _StopSong();
-	void _RegisterSong(byte* data, size_t length);
-	void _UnregisterSong();
-};
-#endif	// OSX
 
 // ============================================================================
 //
@@ -216,7 +169,7 @@ private:
 
 class MidiMusicSystem : public MusicSystem
 {
-public:
+  public:
 	MidiMusicSystem();
 	virtual ~MidiMusicSystem();
 
@@ -224,77 +177,52 @@ public:
 	virtual void stopSong();
 	virtual void pauseSong();
 	virtual void resumeSong();
-	
+
 	virtual void playChunk();
 	virtual void setVolume(float volume);
-	
+
 	// Only plays midi-type music
-	virtual bool isMusCapable() const { return true; }
-	virtual bool isMidiCapable() const { return true; }
-	
-	virtual void playEvent(MidiEvent *event, int time = 0) = 0;
-	
-protected:
+	virtual bool isMusCapable() const
+	{
+		return true;
+	}
+	virtual bool isMidiCapable() const
+	{
+		return true;
+	}
+
+	virtual void playEvent(MidiEvent* event, int time = 0) = 0;
+
+  protected:
 	void _StopSong();
-	
+
 	virtual void _AllNotesOff();
-	
-	int _GetNumChannels() const { return cNumChannels; }
+
+	int _GetNumChannels() const
+	{
+		return cNumChannels;
+	}
 	void _SetChannelVolume(int channel, int volume);
 	void _RefreshVolume();
-	
-	unsigned int _GetLastEventTime() const { return mLastEventTime; }
-	
+
+	unsigned int _GetLastEventTime() const
+	{
+		return mLastEventTime;
+	}
+
 	void _InitializePlayback();
 
 	float _GetScaledVolume();
-	
-private:
-	static const int			cNumChannels = 16;
-	MidiSong*					mMidiSong;
-	MidiSong::const_iterator	mSongItr;
-	bool						mLoop;
-	int							mTimeDivision;
-	
-	unsigned int				mLastEventTime;
-	int							mPrevClockTime;
-	
-	byte						mChannelVolume[cNumChannels];
+
+  private:
+	static const int cNumChannels = 16;
+	MidiSong* mMidiSong;
+	MidiSong::const_iterator mSongItr;
+	bool mLoop;
+	int mTimeDivision;
+
+	unsigned int mLastEventTime;
+	int mPrevClockTime;
+
+	byte mChannelVolume[cNumChannels];
 };
-
-
-// ============================================================================
-//
-// PortMidiMusicSystem class
-//
-// Plays music utilizing the PortMidi music library.
-//
-// ============================================================================
-
-#ifdef PORTMIDI
-class PortMidiMusicSystem : public MidiMusicSystem
-{
-public:
-	PortMidiMusicSystem();
-	virtual ~PortMidiMusicSystem();
-	
-	virtual void stopSong();
-	virtual bool isInitialized() const { return mIsInitialized; }
-		
-	virtual void playEvent(MidiEvent *event, int time = 0);
-
-private:
-	static const int cLatency = 80;
-	
-	bool		mIsInitialized;
-
-	PmDeviceID	mOutputDevice;
-	PmStream*	mStream;
-	
-	void _PlayEvent(MidiEvent *event, int time = 0);
-	void _StopSong();
-};
-#endif	// PORTMIDI
-
-#endif	// __I_MUSICSYSTEM_H__
-
